Create setVideoMode.brs
Writing a plugin that sets the BrightSign videoMode.

Include a function in the plugin that is named:

<plugin name>_SetVideoMode.

This is a similar naming convention to the function that processes events that is named:
<plugin name>_ProcessEvent.

One parameter is supplied to <plugin name>_SetVideoMode.

It is an associative array with the following members:
signVideoMode$ - a string with the video mode as set in the presentation
setPreferredResolution - a boolean that indicates whether or not the autorun will append ":preferred" to the video mode, derived from the Force Resolution setting for the presentation.
fullResGraphicsEnabled - a boolean that indicates the setting for full resolution graphics in the presentation.
tenBitColorEnabled - a boolean that indicates the setting for ten bit color enabled in the presentation.
dolbyVisionEnabled - a boolean that indicates the setting for dolby vision enabled in the presentation.
videoMode$ - a string that represents the resulting videoMode string that the autorun generated based on the values above.

The function <plugin name>_SetVideoMode uses these values, generates a videoMode string and returns that videoMode string to the autorun. The autorun then uses that updated videoMode for the presentation.